explicit cast
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/coin@1224 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/libcoin/src/Contrasts.c
@@ -188,7 +188,7 @@ void C_ordered_Xfactor
     sumright = 0.0;
     for (int p = 0; p < P; p++) 
         sumright += ExpX[p];
-    sw = sumright;
+    sw = (int) sumright;
                  
     for (int p = 0; p < P; p++) {
         sumleft += ExpX[p];
